Apply BEM To Sass code
Block Element Modifiers"BEM" allow use to write code that is easier to
maintain and organize. With BEM elements are categorized into the
"Block"; a stand alone component in a design e.g .card, the Element; child
elements containin in the block often denoted with "__"eg. .card__header
and, the Modifiers; these are alternate versions of either the block or
elements. they are denoted by "--" eg .card--header, --card.