Introduce Sass Mixins to the code base
Mixins are sass built-ins that help use write code resuable code. They
work like functions in regular programing languages. Mixins are created
with the "@mixin" rule, followed by the Mixin name, and the parameter
list; just like a function. within the body of the Mixin we can the
specify the logic we want to resuable when the Mixin is called/used.

To use a defined Mixin, we must first import it into the partial we
we want to use it in using a namespace i.e @use "../util" as u.
then we use the "@include" rule, followed by the Mixin name and pass in
our argument if reqiured.

Sass also allow use to store values in a object like structure called
a Sass-Map. items within a map are store in key-value pairs, where
the key point to the value; just like regular object in programming.
Using Sass built-ins like the "map-get" function, we can retrive items
from sass-maps.

A combination of Sass-maps and Mixins is very powerful and often use to
to write highly flexible and reusable code.